Moved fetch unicode preference function into utilities folder
